svn6046
rr_gl:
@crash pri renderu sceny s vic jak 65535 objekty

+++ src/LightsprintGL/RendererOfMesh.h
@@ -26,7 +26,7 @@ namespace rr_gl
 
 struct FaceGroupRange
 {
-	unsigned short object; // for internal use by RendererOfScene, ignored by RendererOfMesh
+	unsigned object; // for internal use by RendererOfScene, ignored by RendererOfMesh
 	unsigned short faceGroupFirst;
 	unsigned short faceGroupLast;
 	unsigned triangleFirst;
